requirements.txt
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.1.4
charset-normalizer==3.4.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jsonpatch==1.33
jsonpointer==3.0.0
langchain-core>=1.0.0,<2.0.0
langgraph>=1.0.0
langgraph-checkpoint>=4.0.0
langgraph-prebuilt>=1.0.0
langgraph-sdk==0.3.3
langsmith==0.6.7
numpy==2.4.2
orjson==3.11.6
ormsgpack==1.12.2
packaging==25.0
pandas>=1.4.0,<3
pydantic==2.12.5
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
PyYAML==6.0.3
reportlab>=4.0.0
requests==2.32.5
requests-toolbelt==1.0.0
six==1.17.0
tenacity==9.1.2
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uuid_utils==0.14.0
xxhash==3.6.0
zstandard==0.25.0
deepeval>=0.21.0
litellm>=1.0.0
pytest>=7.0.0
pytest~=8.4.1
# guardrails-ai requires langchain-core<0.4 and conflicts with langgraph 1.x.
# Optional: install in a separate env if you need Guard-based validation (guardrails_safety still works without it).
# guardrails-ai~=0.6.6
streamlit~=1.47.0
python-dotenv~=1.1.1
